Resolve PR issues
