+++ src/app/app-routing.module.ts
@@ -3,10 +3,10 @@ import {RouterModule, Routes} from '@angular/router';
 import {LoginComponent} from './login/login.component';
 import {ProblemListComponent} from './problem-list/problem-list.component';
 import { ExamListComponent } from './exam-list/exam-list.component';
-import { ExamIndexComponent } from './exam-index/exam-index.component';
-import { ExamProblemsComponent } from './exam-problems/exam-problems.component';
-import { ExamSubmissionsComponent } from './exam-submissions/exam-submissions.component';
-import { ExamScoreboardComponent } from './exam-scoreboard/exam-scoreboard.component';
+import { ExamHomeComponent } from './exam/exam-home/exam-home.component';
+import { ExamProblemsComponent } from './exam/exam-problems/exam-problems.component';
+import { ExamSubmissionsComponent } from './exam/exam-submissions/exam-submissions.component';
+import { ExamScoreboardComponent } from './exam/exam-scoreboard/exam-scoreboard.component';
 import {MultiTabsPanelComponent} from './problem-submission-tab-panel/multi-tabs-panel.component';
 import {SubmissionsComponent} from './submissions/submissions.component';
 import {ProblemDescriptionComponent} from './problem-description/problem-description.component';
@@ -17,7 +17,7 @@ const routes: Routes = [
   {path: 'problems', component: ProblemListComponent},
   { path: 'exams', component: ExamListComponent, },
   {
-    path: 'exam/:examId', component: ExamIndexComponent,
+    path: 'exams/:examId', component: ExamHomeComponent,
     children: [
       { path: '', component: ExamProblemsComponent, },
       { path: 'submissions', component: ExamSubmissionsComponent, },
