+++ src/app/app.module.ts
@@ -6,10 +6,10 @@ import {LoginComponent} from './login/login.component';
 import {AppComponent} from './app.component';
 import {ProblemListComponent} from './problem-list/problem-list.component';
 import {ExamListComponent} from './exam-list/exam-list.component';
-import { ExamIndexComponent } from './exam-index/exam-index.component';
-import { ExamProblemsComponent } from './exam-problems/exam-problems.component';
-import { ExamScoreboardComponent } from './exam-scoreboard/exam-scoreboard.component';
-import { ExamSubmissionsComponent } from './exam-submissions/exam-submissions.component';
+import { ExamHomeComponent } from './exam/exam-home/exam-home.component';
+import { ExamProblemsComponent } from './exam/exam-problems/exam-problems.component';
+import { ExamScoreboardComponent } from './exam/exam-scoreboard/exam-scoreboard.component';
+import { ExamSubmissionsComponent } from './exam/exam-submissions/exam-submissions.component';
 import {StubStudentService} from './services/impl/StubStudentService';
 import {ExamService, ProblemService, StudentService, SubmissionService} from './services/Services';
 import {StubExamService} from './services/impl/StubExamService';
@@ -48,7 +48,7 @@ const HOST = 'http://api.judgegirl.beta.pdlab.csie.ntu.edu.tw';
     CodeUploadPanelComponent,
     SubmissionsComponent,
 
-    ExamIndexComponent,
+    ExamHomeComponent,
     ExamProblemsComponent,
     ExamSubmissionsComponent,
     ExamScoreboardComponent,
