+++ src/app/problem-description/problem-description.component.ts
@@ -4,9 +4,7 @@ import {ProblemService} from '../services/Services';
 import {ActivatedRoute} from '@angular/router';
 import {switchMap} from 'rxjs/operators';
 import {Observable} from 'rxjs';
-import * as MarkdownIt from 'markdown-it';
-import MarkdownItKatex from 'markdown-it-katex';
-import * as hljs from 'highlight.js';
+import { initHighlight, parseMarkdown } from 'src/utils/mardownUtils';
 
 @Component({
   selector: 'app-problem-description',
@@ -29,7 +27,7 @@ export class ProblemDescriptionComponent implements OnInit, AfterViewInit {
   @ViewChild('compilationScriptPanel') compilationScriptPanel: ElementRef;
 
   ngOnInit(): void {
-    hljs.initHighlightingOnLoad();
+    initHighlight();
     this.problem$ = this.route.parent.params.pipe(switchMap(params =>
       this.problemService.getProblem(+params.problemId)
     ));
@@ -47,33 +45,10 @@ export class ProblemDescriptionComponent implements OnInit, AfterViewInit {
   }
 
   private renderMarkdown() {
-    const markdownIt = new MarkdownIt(
-      {
-        html: true,
-        linkify: true,
-        typographer: true,
-        highlight(str, lang) {  // TODO: highlight currently not work
-          const langModule = hljs.getLanguage(lang);
-          if (lang && langModule) {
-            try {
-              const res = hljs.highlight(lang, str);
-              console.log(`Render language ${lang} to markdown: ${res.value}`);
-              return res.value;
-            } catch (err) {
-              console.log(err);
-            }
-          }
-          return ''; // use external default escaping
-        }
-      }
-    ).use(MarkdownItKatex);
-    markdownIt.renderer.rules.table_open = () => {
-      return '<table class="table">';
-    };
     this.renderer.setProperty(this.markdownPanel.nativeElement, 'innerHTML',
-      markdownIt.render(this.problem.description));
+      parseMarkdown(this.problem.description));
     this.renderer.setProperty(this.compilationScriptPanel.nativeElement, 'innerHTML',
-      markdownIt.render('## Compilation \n' +
+      parseMarkdown('## Compilation \n' +
         '```sh\n' +
         this.problem.compilation.script + '\n' +
         '```'));
